+++ changelog.md
@@ -1,6 +1,13 @@
 # Change log
 
-# # v0.4.0
+## v0.5.0
+
+* You can specify a direct URL to an image when using the **attach URL** feature.
+* The prompt field now accepts both Shift+Enter and Ctrl+Enter for a new line.
+* Changed "Speak response" to "Read response" in the chat menu, so it now uses "R" as the accelerator on Windows.
+* Typo fix
+
+## v0.4.0
 
 * You can now attach multiple images at once.